Feat : Fixed file structure to separate components and css styles.
Moved input form to it's own component file.
